requirements.txt
flask
flask-cors
torch
torchvision
transformers
timm
numpy
opencv-python
pillow
soundfile
requests
werkzeug
